* Merged 224_character_available_pvp_titles.sql to Aspire_Character_Structure.sql * Implemented Add Percent Modifier: Threat Reduced. It fixes Healing Grace * Threat modifiers are also applied to heals. * Fixed a bug in applying threat modifiers. * Removed HandleProc() call from Heal() to prevent spells proccing extra time * Healing Way * GMs can join BGs that are full. BG joining of GMs isn't announced to players and they don't join the group.
git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@275 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ sql/Aspire_Character_Structure.sql
@@ -96,7 +96,7 @@ CREATE TABLE `characters` (
   `skills` longtext COLLATE utf8_unicode_ci NOT NULL,
   `watched_faction_index` int(10) unsigned NOT NULL DEFAULT '0',
   `selected_pvp_title` int(10) unsigned NOT NULL DEFAULT '0',
-  `available_pvp_titles` int(10) unsigned NOT NULL DEFAULT '0',
+  `available_pvp_titles` bigint(40) unsigned NOT NULL DEFAULT '0',
   `gold` int(10) unsigned NOT NULL DEFAULT '0',
   `ammo_id` int(10) unsigned NOT NULL DEFAULT '0',
   `available_prof_points` tinyint(3) unsigned NOT NULL DEFAULT '0',
